quiet deprecation warning about date formatting
In elasticsearch 7.x the 'y' date format has slightly changed meaning
and emits deprecation warnings. Update to 'u' which has the same
meaning as 'y' used to, along with prefixing the '8' which tells
elastic 6.8 to use the new date formatting code. Once the elastic 7
transition is complete the '8' prefix can be removed.

Tested in elastic 6.8 and 7.10.

Change-Id: I5697fbadcf3d985b095a55a295b1a5f6862a6763