Add option to use a SQL table instead of ElasticSearch
This query engine subclass uses adaptive sampling to heavily
limit the number of writes, which might be cross-datacenter.
The reported numbers are daily estimates. However, the first
daily occurence of an agent/feature is always sampled 1:1.
This is done to capture the presence of clients making
infrequent requests involving deprecated features.

Use the SQL query engine as the default. Previously, an error
was thrown if $wgApiFeatureUsageQueryEngineConf was not set.

Add a ApiLogFeatureUsage hook handler to notify the query engine
when deprecated features are used. Previously, external services
had to be setup to listen to the UDP logs, parse them, and send
updates to the backing-service of the query engine.

Move the Elastica hard dependency to a lazy soft dependency
since there are now two implementations.

Migrate the ApiFeatureUsageQueryEngine to a service class.

Bug: T313731
Depends-On: I8bbf806501d542a0f91a67e197913a894bdee127
Change-Id: Ie10612373ba5375c6a66fae13f65dc1ca98959a6